Move hide module out from Main
