+++ app/src/main/java/me/veryyoung/qq/luckymoney/Main.java
@@ -1,20 +1,13 @@
 package me.veryyoung.qq.luckymoney;
 
 import android.app.Activity;
-import android.app.ActivityManager.RunningAppProcessInfo;
-import android.app.ActivityManager.RunningServiceInfo;
-import android.app.ActivityManager.RunningTaskInfo;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.ApplicationInfo;
-import android.content.pm.PackageInfo;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.text.TextUtils;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 import java.util.Random;
 
@@ -33,11 +26,12 @@ import static de.robv.android.xposed.XposedHelpers.findClass;
 import static de.robv.android.xposed.XposedHelpers.findFirstFieldByExactType;
 import static de.robv.android.xposed.XposedHelpers.getObjectField;
 import static de.robv.android.xposed.XposedHelpers.newInstance;
+import static me.veryyoung.qq.luckymoney.HideModule.hideModule;
 
 
 public class Main implements IXposedHookLoadPackage {
 
-    private static final String QQ_PACKAGE_NAME = "com.tencent.mobileqq";
+    public static final String QQ_PACKAGE_NAME = "com.tencent.mobileqq";
     private static final String WECHAT_PACKAGE_NAME = "com.tencent.mm";
 
 
@@ -217,118 +211,6 @@ public class Main implements IXposedHookLoadPackage {
             });
         }
 
-    }
-
-
-    private void hideModule(XC_LoadPackage.LoadPackageParam loadPackageParam) {
-        findAndHookMethod("android.app.ApplicationPackageManager", loadPackageParam.classLoader, "getInstalledApplications", int.class, new XC_MethodHook() {
-            @Override
-            protected void afterHookedMethod(MethodHookParam param) throws Throwable {
-                List<ApplicationInfo> applicationList = (List) param.getResult();
-                List<ApplicationInfo> resultapplicationList = new ArrayList<>();
-                for (ApplicationInfo applicationInfo : applicationList) {
-                    String packageName = applicationInfo.packageName;
-                    if (isTarget(packageName)) {
-                        log("Hid package: " + packageName);
-                    } else {
-                        resultapplicationList.add(applicationInfo);
-                    }
-                }
-                param.setResult(resultapplicationList);
-            }
-        });
-        findAndHookMethod("android.app.ApplicationPackageManager", loadPackageParam.classLoader, "getInstalledPackages", int.class, new XC_MethodHook() {
-            @Override
-            protected void afterHookedMethod(MethodHookParam param) throws Throwable {
-                List<PackageInfo> packageInfoList = (List) param.getResult();
-                List<PackageInfo> resultpackageInfoList = new ArrayList<>();
-
-                for (PackageInfo packageInfo : packageInfoList) {
-                    String packageName = packageInfo.packageName;
-                    if (isTarget(packageName)) {
-                        log("Hid package: " + packageName);
-                    } else {
-                        resultpackageInfoList.add(packageInfo);
-                    }
-                }
-                param.setResult(resultpackageInfoList);
-            }
-        });
-        findAndHookMethod("android.app.ApplicationPackageManager", loadPackageParam.classLoader, "getPackageInfo", String.class, int.class, new XC_MethodHook() {
-            @Override
-            protected void beforeHookedMethod(MethodHookParam param) throws Throwable {
-                String packageName = (String) param.args[0];
-                if (isTarget(packageName)) {
-                    param.args[0] = QQ_PACKAGE_NAME;
-                    log("Fake package: " + packageName + " as " + QQ_PACKAGE_NAME);
-                }
-            }
-        });
-        findAndHookMethod("android.app.ApplicationPackageManager", loadPackageParam.classLoader, "getApplicationInfo", String.class, int.class, new XC_MethodHook() {
-            @Override
-            protected void beforeHookedMethod(MethodHookParam param) throws Throwable {
-                String packageName = (String) param.args[0];
-                if (isTarget(packageName)) {
-                    param.args[0] = QQ_PACKAGE_NAME;
-                    log("Fake package: " + packageName + " as " + QQ_PACKAGE_NAME);
-                }
-            }
-        });
-        findAndHookMethod("android.app.ActivityManager", loadPackageParam.classLoader, "getRunningServices", int.class, new XC_MethodHook() {
-            @Override
-            protected void afterHookedMethod(MethodHookParam param) throws Throwable {
-                List<RunningServiceInfo> serviceInfoList = (List) param.getResult();
-                List<RunningServiceInfo> resultList = new ArrayList<>();
-
-                for (RunningServiceInfo runningServiceInfo : serviceInfoList) {
-                    String serviceName = runningServiceInfo.process;
-                    if (isTarget(serviceName)) {
-                        log("Hid service: " + serviceName);
-                    } else {
-                        resultList.add(runningServiceInfo);
-                    }
-                }
-                param.setResult(resultList);
-            }
-        });
-        findAndHookMethod("android.app.ActivityManager", loadPackageParam.classLoader, "getRunningTasks", int.class, new XC_MethodHook() {
-            @Override
-            protected void afterHookedMethod(MethodHookParam param) throws Throwable {
-                List<RunningTaskInfo> serviceInfoList = (List) param.getResult();
-                List<RunningTaskInfo> resultList = new ArrayList<>();
-
-                for (RunningTaskInfo runningTaskInfo : serviceInfoList) {
-                    String taskName = runningTaskInfo.baseActivity.flattenToString();
-                    if (isTarget(taskName)) {
-                        log("Hid task: " + taskName);
-                    } else {
-                        resultList.add(runningTaskInfo);
-                    }
-                }
-                param.setResult(resultList);
-            }
-        });
-        findAndHookMethod("android.app.ActivityManager", loadPackageParam.classLoader, "getRunningAppProcesses", new XC_MethodHook() {
-            @Override
-            protected void afterHookedMethod(MethodHookParam param) throws Throwable {
-                List<RunningAppProcessInfo> runningAppProcessInfos = (List) param.getResult();
-                List<RunningAppProcessInfo> resultList = new ArrayList<>();
-
-                for (RunningAppProcessInfo runningAppProcessInfo : runningAppProcessInfos) {
-                    String processName = runningAppProcessInfo.processName;
-                    if (isTarget(processName)) {
-                        log("Hid process: " + processName);
-                    } else {
-                        resultList.add(runningAppProcessInfo);
-                    }
-                }
-                param.setResult(resultList);
-            }
-        });
-    }
-
-    private boolean isTarget(String name) {
-        return name.contains("veryyoung") || name.contains("xposed");
     }
