+++ gradle.properties
@@ -5,15 +5,15 @@
 #
 # Specifies the JVM arguments used for the daemon process.
 # The setting is particularly useful for tweaking memory settings.
-# Default value: -Xmx10248m -XX:MaxPermSize=256m
+# Default value: -Xmx1024m -XX:MaxPermSize=256m
 # org.gradle.jvmargs=-Xmx2048m -XX:MaxPermSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
 #
 # When configured, Gradle will run in incubating parallel mode.
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
-#Sun Nov 29 01:38:09 CST 2015
-systemProp.https.proxyPort=1080
+#Sun Dec 11 21:11:41 CST 2016
 systemProp.http.proxyHost=127.0.0.1
+systemProp.https.proxyPort=1080
 systemProp.https.proxyHost=127.0.0.1
-systemProp.http.proxyPort=1080
+systemProp.http.proxyPort=8123
